test: Automated tests improvements for Mongo CRUD cases + few flaky fix (#11985)
* Adding AForceMigration spec

* Merge from release

* Mongo CRUD data load check added

* Added bug id & adding to skip to checkin

* Unskipping MySQL_Spec to test TED

* Api-bugs spec fix

* Adding back MySQL_Spec to ignore

* EE_CopyQuery_RenameDatasource_spec revert to fix

* Skipping Git_sync flaky test

* Removing mySQl from ignore

* Unskipping Mongo cases as Bug12162 fixed

* Empty files moved due to Mocha update
