+++ app/client/cypress/integration/Smoke_TestSuite/ServerSideTests/ExplorerTests/EditorContextMenu_Spec.ts
@@ -59,7 +59,7 @@ describe("Validate basic operations on Entity explorer JSEditor structure", () =
 
   it("6. Validate Deletion of JSObject", function() {
     agHelper.SelectEntityByName(pageId);
-    cy.CheckAndUnfoldEntityItem("QUERIES/JS");
+    agHelper.expandCollapseEntity("QUERIES/JS");
     agHelper.ActionContextMenuByEntityName(
       "ExplorerRenamed",
       "Delete",
