+++ app/client/cypress/manual_TestSuite/CommentedScriptFiles/AForceMigration_Spec.ts
@@ -1,5 +1,5 @@
 // const widgetsPage = require("../../../locators/Widgets.json");
-// import homePage from "../../../locators/HomePage";
+// import homePage from "../../locators/HomePage";
 
 // describe("Migration Validate", function () {
 //   it("1. Import application and Validate Migration on pageload", function () {
